Moved database structs & sqlite initialization to DBSchema.h
